fix: UAF caused by C++ shenanigans
The problem here was, I believe, that "auto" inferred the wrong type
for "bytes": it was a Vector<uint8_t>, not a reference to a const
Vector<uint8_t>. The problem is that the former has a destructor which
frees the buffer, but we wanted to steal the buffer and destroy the
vector way later when its owning object would be destroyed.

Making the type explicit seems safest here. "const auto&" would also
work.